Split layerLabel to more files, export clickOutside hook, set layerLabel height auto
