+++ src/components/ExpandableLabel/style.scss
@@ -72,15 +72,21 @@ $ExpandableLabelRoundedBorderRadius: 1.5rem;
 
 // Body
 .ptr-ExpandableLabelBody {
-	min-width: 10rem;
 	margin-top: m(math.div(1, 3));
-	height: 0;
 	overflow-y: hidden;
-	transition: height 0.3s ease-in-out;
 	border-radius: $ExpandableLabelBorderRadius;
 	background: var(--base05);
 	position: absolute;
 	z-index: 2;
+	height: 0px;
+	transition: height 0.3s ease-in-out;
+
+	& > div {
+		height: auto;
+		overflow-y: hidden;
+		border: solid 1px var(--base30);
+		border-radius: .25rem;
+	}
 }
 
 .ptr-ExpandableLabel.is-floating .ptr-ExpandableLabelBody {
@@ -96,16 +102,11 @@ $ExpandableLabelRoundedBorderRadius: 1.5rem;
 	transform: rotate(90deg);
 }
 
-.ptr-ExpandableLabelBody.is-expanded {
-	display: flex;
-	height: 10rem;
-}
-
 .ptr-dark .ptr-ExpandableLabelBody.is-expanded {
 	border: 1px solid var(--base30);
 }
 
 .ptr-ExpandableLabelBody > div {
-	height: 10rem;
+	// height: 10rem;
 	overflow-y: auto;
 }
